UI 수정
 - 기록에 얼리 레이트 개수를 표시하도록 추가.

 - 결과 화면에서 입력 타이밍 그래프 Kool 기준선 위치 조정, 점 크기를 축소하여 더 정확하게 표시하도록 수정.

+++ D-BMS/Assets/Scripts/BMSParser.cs
@@ -343,7 +343,7 @@ public class BMSParser : MonoBehaviour
 
         double beat16Count = 16 * pattern.GetBeatC(currentBar);
 
-        if (currentBar == bar && prevBeat >= endBeat) { return; }
+        if (currentBar == bar && (float)prevBeat >= (float)endBeat) { return; }
 
         while (true)
         {
@@ -361,7 +361,7 @@ public class BMSParser : MonoBehaviour
                 prevBeat = -4.0d;
             }
 
-            if ((currentBar == bar && currentBeat > endBeat) || currentBar > bar) { break; }
+            if ((currentBar == bar && (float)currentBeat > (float)endBeat) || currentBar > bar) { break; }
 
             pattern.AddNote(line, currentBar, currentBeat, 0, 2);
         }
